// src/Moryx.Orders.Web/app/src/app/api/models/begin-context.ts
/*
 * Copyright (c) 2026 Phoenix Contact GmbH & Co. KG
 * Licensed under the Apache License, Version 2.0
*/

/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { RestrictionDescription } from '../models/restriction-description';
export interface BeginContext {
  canBegin?: boolean;
  canReduce?: boolean;
  minimalTargetAmount?: number;
  partialAmount?: number;
  residualAmount?: number;
  restrictions?: Array<RestrictionDescription> | null;
  scrapCount?: number;
  successCount?: number;
}